refactor(build): migrate tsup → tsdown across all packages
tsup 8.5.1's DTS plugin unconditionally injects `baseUrl: "."` into the
compiler options it hands to rollup-plugin-dts. TypeScript 6.0 rejects
that with TS5101, blocking the root TS bump (see #2491). The workaround
in tests.yml only patched the test-types matrix and not the build job —
piling more workarounds on a deprecated bundler is the wrong direction.

tsdown is rolldown-based, uses rolldown-plugin-dts (no baseUrl
injection), declares peer typescript ^5 || ^6, and tsup's own README
recommends it.

What changes:

- tsup.config.base.ts → tsdown.config.base.mjs (+ matching .d.mts). JS +
  JSDoc rather than TS so the per-package tsdown.config.ts files load it
  through the native ESM loader; pnpm's strict isolation collides with
  tsdown's tsx loader.
- All 12 per-package tsup.config.ts → tsdown.config.ts.
- Each package's build/dev scripts swap tsup → tsdown and pre-clean
  dist/ (tsdown's clean defaults differ).
- Drop the now-dead test-types '6.0' jq patch + comment block.

The six-bundle layout (cjs dev + prod-min and esm legacy + modern +
modern.dev + modern.prod-min) is preserved byte-for-byte at the file
listing level — packages' public `exports` fields are unchanged.

Known behaviour difference: for the modern.mjs entry (the one carrying
.d.mts), rolldown extracts its `__exportAll` runtime helper into a
sibling dist/chunk.mjs that codeSplitting:false does not suppress in
tsdown 0.22 / rolldown 1.0. The chunk ships via each package's
`files: ["dist/**/*"]` and is resolved as a relative import from the
entry, so consumers are unaffected; are-the-types-wrong is green across
all resolution modes for the produced tarballs.

Refs: egoist/tsup#1388.
