+++ package.json
@@ -85,7 +85,7 @@
     "react-zdog": "1.2.2",
     "spec.ts": "1.1.3",
     "three": "0.184.0",
-    "tsup": "8.5.1",
+    "tsdown": "0.22.0",
     "turbo": "2.9.14",
     "typescript": "5.7.2",
     "vite": "8.0.14",
