+++ packages/animated/package.json
@@ -48,9 +48,9 @@
     "provenance": true
   },
   "scripts": {
-    "build": "tsup",
+    "build": "rm -rf dist && tsdown",
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
-    "dev": "tsup --watch",
+    "dev": "tsdown --watch",
     "pack": "pnpm pack && mv *.tgz package.tgz"
   },
   "dependencies": {
