+++ packages/core/package.json
@@ -51,9 +51,9 @@
     "provenance": true
   },
   "scripts": {
-    "build": "tsup",
+    "build": "rm -rf dist && tsdown",
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
-    "dev": "tsup --watch",
+    "dev": "tsdown --watch",
     "pack": "pnpm pack && mv *.tgz package.tgz"
   },
   "dependencies": {
