--- packages/core/tsup.config.ts
@@ -1,12 +1,0 @@
-import { defineConfig } from 'tsup'
-import { defaultConfig } from '../../tsup.config.base'
-
-export default defineConfig(opts => {
-  return defaultConfig(
-    {
-      entry: 'src/index.ts',
-      name: 'react-spring_core',
-    },
-    opts
-  )
-})
