+++ targets/konva/package.json
@@ -58,9 +58,9 @@
     "react-konva": "^19"
   },
   "scripts": {
-    "build": "tsup",
+    "build": "rm -rf dist && tsdown",
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
-    "dev": "tsup --watch",
+    "dev": "tsdown --watch",
     "pack": "pnpm pack && mv *.tgz package.tgz"
   }
 }
