+++ targets/konva/tsdown.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from 'tsdown'
+import { defaultConfig } from '../../tsdown.config.base.mjs'
+
+export default defineConfig(
+  defaultConfig({
+    entry: 'src/index.ts',
+    name: 'react-spring_konva',
+  })
+)
