+++ targets/native/tsdown.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from 'tsdown'
+import { defaultConfig } from '../../tsdown.config.base.mjs'
+
+export default defineConfig(
+  defaultConfig({
+    entry: 'src/index.ts',
+    name: 'react-spring_native',
+    buildFilter: ['development', 'production.min'],
+  })
+)
