--- targets/native/tsup.config.ts
@@ -1,13 +1,0 @@
-import { defineConfig } from 'tsup'
-import { defaultConfig } from '../../tsup.config.base'
-
-export default defineConfig(opts => {
-  return defaultConfig(
-    {
-      entry: 'src/index.ts',
-      name: 'react-spring_native',
-      buildFilter: ['development', 'production.min'],
-    },
-    opts
-  )
-})
