+++ targets/zdog/package.json
@@ -59,9 +59,9 @@
     "zdog": ">=1.0"
   },
   "scripts": {
-    "build": "tsup",
+    "build": "rm -rf dist && tsdown",
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
-    "dev": "tsup --watch",
+    "dev": "tsdown --watch",
     "pack": "pnpm pack && mv *.tgz package.tgz"
   }
 }
